requirements.txt
fastapi>=0.101.1,<1.0.0
gunicorn>=20.1.0,<21.0.0
uvicorn>=0.20.0,<1.0.0
google-cloud-logging>=3.5.0,<4.0.0
pydantic>=1.10.7,<2.0.0
google-cloud-firestore >=2.3.4
mock-firestore-async>=0.11.3
httpx>=0.24.0,<1.0.0
trio>=0.22.2
pytest>=7.4.0
anyio>=3.7.1